Specify required Visual Studio component
Before this change, it was possible for users to get stuck during step 0
of the build instructions. Specifically, step 0 told you to run this
command:

	"%ProgramFiles(x86)%\Microsoft Visual Studio\Installer\setup.exe"

That command simply opens the Visual Studio Installer. The build
instructions did not tell you what to do once the Visual Studio
Installer is opened. The instructions do say to “install your preferred
workflow but be sure to include the vcpkg package manager”, but they
don’t explain to the user how to install the vcpkg package manager.

This change modifies the command so that it ensures that the vcpkg
package manager is installed without requiring user interaction. That
way, users won’t get stuck on step 0.