Tell user to use an Administrator Command Prompt
If you tried to follow the build instructions before this change, then
there’s a chance that you would get an error. Specifically, users might
run this command during step 0 of the build instructions:

    "%ProgramFiles(x86)%\Microsoft Visual Studio\Installer\setup.exe" modify^
        --passive^
        --channelId VisualStudio.17.Release^
        --productId Microsoft.VisualStudio.Product.BuildTools^
        --add Microsoft.VisualStudio.Component.Vcpkg

If that command is run in a regular (non-Administrator) Command Prompt,
then it will fail with this error:

> [3adc:0001][2025-04-12T15:58:08] Commands with --quiet or --passive should be run elevated from the beginning.
> [3adc:0001][2025-04-12T15:58:08] Exit Code: 5007

This change tells users to run the command in an Administrator Command
Prompt in order to avoid that error. A side effect of this change is
that users are now told to use an Administrator Command Prompt when
running a winget command as well. In this particular scenario, it’s
probably not necessary to run winget in an Administrator Command Prompt,
but I generally prefer running winget in an Administrator Command
Prompt. Some winget commands need to be run in an Administrator Command
Prompt, and running winget commands in an Administrator Command Prompt
allows you to respond to a UAC prompt before any installers have been
downloaded.