Use VS Community instead of VS Build Tools
Before this change, there was a chance that users would not be able to
perform step 2 of the build instructions. Steps to reproduce that
situation:

1. Make sure that the Visual Studio Installer is not installed.
2. Try to follow this repo’s build instructions. You’ll get to a point
   where you need to open the Visual Studio IDE, but you won’t be able
   to because it won’t be installed.

The failure would happen because there was no guarantee that users had
actually installed the Visual Studio IDE. There are several different
editions of Visual Studio 2022. Confusingly, Visual Studio Build Tools
2022 is an edition of Visual Studio 2022 that doesn’t include the Visual
Studio IDE [1].

This change prevents the failure by telling users to install a different
edition of Visual Studio 2022. Visual Studio Community 2022 was chosen
because it doesn’t cost anything and (as far as I know) it’s the most
basic edition of Visual Studio that includes the Visual Studio IDE.

[1]: <https://learn.microsoft.com/en-us/visualstudio/install/workload-and-component-ids?view=vs-2022>