Tell users to install the Desktop development with C++ workload
Before this change, there was a chance that the build instructions would
fail. Steps to reproduce that situation:

1. Make sure that the Visual Studio Installer is not installed.
2. Try to follow this repo’s build instructions. You’ll get to a point
   where you need to run a command in a Developer Command Prompt. When
   you try to run that command, it will fail with this error message:

   > error: in triplet x64-windows: Unable to find a valid Visual Studio instance
   > Could not locate a complete Visual Studio instance
   > The following paths were examined for Visual Studio instances:
   >   C:\Program Files\Microsoft Visual Studio\2022\Community\VC\Auxiliary/Build\vcvarsall.bat
   >   C:\Program Files\Microsoft Visual Studio\2022\Community\VC\Auxiliary/Build\vcvarsall.bat

This change prevents those errors from happening.

This change also removes this line from the build instructions’s Visual
Studio Installer command:

>      --add Microsoft.VisualStudio.Component.Vcpkg

That line is no longer needed because we’re now installing the Desktop
development with C++ workload. vcpkg is one of the recommended
components of that workload.